Remove all file ignores, split common
