+++ src/ephys_link/__main__.py
@@ -15,17 +15,14 @@ from ephys_link.back_end.server import Server
 from ephys_link.front_end.cli import CLI
 from ephys_link.front_end.gui import GUI
 from ephys_link.utils.console import Console
+from ephys_link.utils.startup import check_for_updates, preamble
 
 
 def main() -> None:
-    """Ephys Link entry point.
+    """Ephys Link entry point."""
 
-    1. Get options via CLI or GUI.
-    2. Instantiate the Console and make it globally accessible.
-    3. Instantiate the Platform Handler with the appropriate platform bindings.
-    4. Instantiate the Emergency Stop service.
-    5. Start the server.
-    """
+    # 0. Print the startup preamble.
+    preamble()
 
     # 1. Get options via CLI or GUI (if no CLI options are provided).
     options = CLI().parse_args() if len(argv) > 1 else GUI().get_options()
@@ -33,12 +30,16 @@ def main() -> None:
     # 2. Instantiate the Console and make it globally accessible.
     console = Console(enable_debug=options.debug)
 
-    # 3. Instantiate the Platform Handler with the appropriate platform bindings.
+    # 3. Check for updates if not disabled.
+    if not options.ignore_updates:
+        check_for_updates(console)
+
+    # 4. Instantiate the Platform Handler with the appropriate platform bindings.
     platform_handler = PlatformHandler(options, console)
 
-    # 4. Instantiate the Emergency Stop service.
+    # 5. Instantiate the Emergency Stop service.
 
-    # 5. Start the server.
+    # 6. Start the server.
     Server(options, platform_handler, console).launch()
